Drain warning flood during compilation
First we disable the unused-parameter warning, because that
is raised so often, it is counterproductive to show.

There were also some instances where we checked for !this in
methods and some instances where an int was assigned NULL.
Those are now refactored.

This commit also contains some code cleanups (whitespace and
stuff along the way).