Autodetect source files; avoid copy&paste
Before this commit all the files to compile had to be added
to both the client's AND the server's list of files to
compile.

(This commit is a bit monolithic, sorry for that).

Specifically the following changes have been made:
* Most headers and sources are now included by searching for
  them in the build process as opposed to manually listing them
* This also reflects the modules approach we've been taking
  in respecto to namespaces and folders.
* Configuration specific to a module has been moved into
  that modules directory.
* declare_module(...) has been added to search for files.
* prepend(...) has been added to ease listing source files
* create_names(...) has been removed in favor of prepend(...)

Modules & Auto detection
========================

This was not very safe and a bit of effort. Beginning with
this commit, most files are added automatically:

We can now declare build modules, these are added by using
a special CMake function.
Modules consinst of a name and a directory. The directory is
searched for any *.cpp or *.h files. These files are
automatically added to the corresponding VS groups and the
lists are exported in variables.
These lists can be used in the targets to select specific
modules for client and server.

The configuration specific to a module have been moved to
that module:
* rpc/ and protobuf -> rpc/
* net/ and ASIO -> net/
* muparser -> libraries/

protoc compiler
---------------

To really keep the configuration of the protobuf protoc/
local to rpc/, it now exports a function that does all the
scope-local or target-specific functions inside client/ and
server/.

muparser
--------

This is not possible in some cases: Muparser has files in
it's source folder we can't compile into our binary, hence
we manually maintain a *single* list of files in
src/libraries/.

common configuration
====================

Libs, headers, and source files that are used in all targets
are now saved in central variables in src/.

fpsgame/ and engine/
====================

fpsgame/ and engine/ are detected as modules, the client and
the server however require distinct lists of files to
compile. For those folders we maintain separate lists inside
client/ and server/.

Since adding the proper groups is already handled by the
module system, we no longer need to create the groups in
those list. As a consequence of this we no longer use
create_names to canonicalize the paths of source files; now
we use a simple prepend(...) function to generate those
paths.