Auto detect headers/libs; improve dir structure
Note: This commit has a corresponding commit in the
platform_windows repository

Summary
=======

This is a complex commit. Sorry.

* We now use a variable INSTALL_SHARED_LIBS to name the
dynlibs to install into the executable dir. This allows us
to get rid of platform_windows/CMakeLists.txt
* Use a new naming scheme for the lib directory based on
  PLATFORM-ARCH-BUILD_CONFIG
* Move all the shared libs from platform_windows/bin/** into
  platform_windows/lib/**, so mingw can find them
* Detect the path to the protoc compiler during running
  cmake. This way setting PATH takes effect.

Why?
====

On most platforms we are using find_library and find_path to
detect where libs/headers are.

On windows though, before this commit we were using
a CMakeLists.txt file inside platform_windows which would
set the locations of libs/header dirs (these are the
variables like SDL_INCLUDE_DIR or SDL_LIBRARIES).
This was bad, because:
* Inexor's CMakeLists.txt needed knowledge of the
  platform_windows directory. (Which was hard coded)
* In general, it should be avoided to place platform
  specific configuration outside the toolchain file
* There were multiple places that would declare a libraries
  files/headers. platform_windows and the detection functions)
* Deactivating the detection functions on win prevents users
  from supplying their own toolchain/libs

In order to improve this situation, libs on windows are now
also subject to the auto detection: We simply add
platform_windows/include/ to the header search path and the
appropriate targets from platform_windows/lib to the library
search path.