cmake: Improve module encapsulation
NOTE: By saying module I mean net, util and rpc. For the
build system muparser is also treated as a module.

Before this commit some parts of a modules (net, util, rpc)
configuration would leak into the executable target's
configuration, because some things could only be state
there.
In addition to that we would compile modules by just adding
their source files to the executable target's source files.
This had the additional drawback, that we had to build each
of the module's sources once for each executable target
(which adds about 8 source files  per executable at the
moment).

In oder to tackle this problem, this commit introduces two
changes:

1) We are now compiling each module into a static library.
That library is then linked into the actual executable.

2) Each module now provides a config_<NAME> function.
Calling this function with a target must be the sole action
required to add a module to an executable target.
These functions add their associated library to the target,
they might add external libraries (e.g. protobuf, asio) to
the target, the might add additional file creation targets
or add additional include directories.
Since they are functions, called from within the executable
target's scope, they are almost as powerful as writing code
in the file itself – with the exception of code that must be
executed before the target is created (e.g.
link_directories())