cmake: Fix headers missing in VS groups; easier lib/module declaration
Before this commit, some headers had – once again – not been
shown in the VS source groups. This was fixed by adding the
headers to the targets. (Fuck you CMake)

We also add a add_lib module that simplifies the
instantiation of module libs like rpc, net, util, ...

In addition to that there are some indentation cleanups  and
we rename target_link_libraries_wrapper to target_link_libs