Fix a memory leak in the Socket Message Connect code
In the function that accepts new connections for the Socket
Message Connect server, we created an Message Connect
pointer, which would later be filed with an actual
connection.
If, however, no connection was accepted, this pointer was
not deleted.

In addition to that, if an exception would happen in that
code, it would also create a mem leak.

We solve this by using a unique_ptr instead of a plain
pointer, so the pointer is automatically freed unless it is
moved outside the scope.

This commit also pulls some other changes:
* The MCHub was updated to deal with not only pointer, but
  also pointer-like types in the list of MCs
* We provide broadcast/receiveOnAny functions that use
  collections or iterators as hub, without needing to create
  a hub object
* In many cases we now use constant function parameters