Fix inter-class state contagion between Injectables
First: Before this commit all Injectables woudl use the same
list of dependencies, because we stored the list ass a class
variable of Injectable, which was then shared among the
subclasses.
Instead we now default-initialize the variable the first
time it is accessed (in @inject). Since @inject is called
from the subclass, the variable will be registered with the
subclass.

Second: We no longer store a @clz instance variable in
Injectable. @clz was supposed to be a fast way to access the
own class, but didn't quite work and we already have that
variable: @constructor and @constructor works perfect.