[metasystem] Defer subsystem starts until the next tick
This adds a generic facility to metasystem that can be used
to register one time callbacks for the next tick and uses
that facility to defer starting/stopping subsystems using
that.

start() and stop() are modified, so they don't throw any
errors when starting/stopping an already stared/stopped
subsystem.