Fix key being printed in chat
In the startup procedure of the engine SDL_StartTextInput() was called.
That enabled the SDL_TEXTINPUT events that are needed to type text in
chat. The problem with that is that SDL_TEXTINPUT events are not needed
unless one types text, which is the case right at the beginning.

Fixes #279