bug-workaround: cubescript only files in seperate modules do not work
so we merge texture/legacyslotload.cpp and texture/slot.cpp (since thats a logical unit anyways)

This bug occurs when you have only "unused" code (cubescript) in a file in a module.

It's caused by the internal architecture of inexor, where modules are static libs (so we dont need to compile everything multitple times for different applications).
So those files get ignored by the compiler completely.

As soon as anything in those files get referenced the bug disappears.