[GLUEGEN] working parsing of Shared Class Singletons!
This is the first working version of SharedClass parsing.
We detect classes which derived from SharedClass and any instance of those (which must be a non-ptr type to work correctly currently) gets automatically added and rendered.

Currently nested classes are not working and the namespaces of the instances are not properly checked.