feat(cmake): use conan targets instead of libs
the problem we had before is that conan lists the libs relative to the conan library dirs, while if it fails to find such a relative lib, it will look into the system dirs.
If it finds such a system lib, it will link to it, while it should error instead. Otherwise we do not know something weird is happening behind the scenes.

We abandon the relative libs and switch to using cmake interface targets. Those bundle all dependencies of the required lib, all linker and compiler flags.. everything into one single virtual target.